Re-arrange code to avoid evaluating creds in test
