+++ sipn_reanalysis_ingest/util/download.py
@@ -8,24 +8,8 @@ from loguru import logger
 from sipn_reanalysis_ingest._types import CfsrProductType
 from sipn_reanalysis_ingest.constants.creds import RDA_PASSWORD, RDA_USER
 from sipn_reanalysis_ingest.constants.download import DOWNLOAD_AUTH_URL
-from sipn_reanalysis_ingest.constants.paths import DATA_DOWNLOAD_DIR
 from sipn_reanalysis_ingest.errors import CredentialsError, DownloadError
-from sipn_reanalysis_ingest.util.cfsr import cfsr_5day_input_identifier
 from sipn_reanalysis_ingest.util.url import cfsr_5day_tar_url
-
-
-def download_dir(
-    *,
-    window_start: dt.date,
-    window_end: dt.date,
-    product_type: CfsrProductType,
-) -> Path:
-    subdir_name = cfsr_5day_input_identifier(
-        window_start=window_start,
-        window_end=window_end,
-        product_type=product_type,
-    )
-    return DATA_DOWNLOAD_DIR / subdir_name
 
 
 @cache
